requirements.txt
aiofiles==24.1.0 ; python_version == "3.12"
aiohappyeyeballs==2.6.1 ; python_version == "3.12"
aiohttp==3.11.18 ; python_version == "3.12"
aiosignal==1.3.2 ; python_version == "3.12"
altair==5.5.0 ; python_version == "3.12"
annotated-types==0.7.0 ; python_version == "3.12"
anyio==4.9.0 ; python_version == "3.12"
attrs==25.3.0 ; python_version == "3.12"
backoff==2.2.1 ; python_version == "3.12"
beautifulsoup4==4.13.4 ; python_version == "3.12"
blinker==1.9.0 ; python_version == "3.12"
cachetools==5.5.2 ; python_version == "3.12"
certifi==2025.4.26 ; python_version == "3.12"
cffi==1.17.1 ; python_version == "3.12"
chardet==5.2.0 ; python_version == "3.12"
charset-normalizer==3.4.1 ; python_version == "3.12"
click==8.1.8 ; python_version == "3.12"
colorama==0.4.6 ; python_version == "3.12" and platform_system == "Windows"
contourpy==1.3.2 ; python_version == "3.12"
cryptography==44.0.2 ; python_version == "3.12"
cycler==0.12.1 ; python_version == "3.12"
dataclasses-json==0.6.7 ; python_version == "3.12"
deprecated==1.2.18 ; python_version == "3.12"
distro==1.9.0 ; python_version == "3.12"
emoji==2.14.1 ; python_version == "3.12"
et-xmlfile==2.0.0 ; python_version == "3.12"
eval-type-backport==0.2.2 ; python_version == "3.12"
filetype==1.2.0 ; python_version == "3.12"
fonttools==4.57.0 ; python_version == "3.12"
frozenlist==1.6.0 ; python_version == "3.12"
gitdb==4.0.12 ; python_version == "3.12"
gitpython==3.1.44 ; python_version == "3.12"
google-api-core==2.24.2 ; python_version == "3.12"
google-auth==2.39.0 ; python_version == "3.12"
google-cloud-appengine-logging==1.6.1 ; python_version == "3.12"
google-cloud-audit-log==0.3.2 ; python_version == "3.12"
google-cloud-core==2.4.3 ; python_version == "3.12"
google-cloud-logging==3.12.1 ; python_version == "3.12"
googleapis-common-protos==1.70.0 ; python_version == "3.12"
greenlet==3.2.1 ; python_version == "3.12" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32")
grpc-google-iam-v1==0.14.2 ; python_version == "3.12"
grpcio-status==1.71.0 ; python_version == "3.12"
grpcio==1.71.0 ; python_version == "3.12"
h11==0.16.0 ; python_version == "3.12"
httpcore==1.0.9 ; python_version == "3.12"
httpx-sse==0.4.0 ; python_version == "3.12"
httpx==0.28.1 ; python_version == "3.12"
idna==3.10 ; python_version == "3.12"
importlib-metadata==8.6.1 ; python_version == "3.12"
jinja2==3.1.6 ; python_version == "3.12"
jiter==0.9.0 ; python_version == "3.12"
joblib==1.4.2 ; python_version == "3.12"
jsonpatch==1.33 ; python_version == "3.12"
jsonpointer==3.0.0 ; python_version == "3.12"
jsonschema-specifications==2025.4.1 ; python_version == "3.12"
jsonschema==4.23.0 ; python_version == "3.12"
kiwisolver==1.4.8 ; python_version == "3.12"
langchain-community==0.3.23 ; python_version == "3.12"
langchain-core==0.3.56 ; python_version == "3.12"
langchain-experimental==0.3.4 ; python_version == "3.12"
langchain-openai==0.2.14 ; python_version == "3.12"
langchain-text-splitters==0.3.8 ; python_version == "3.12"
langchain==0.3.24 ; python_version == "3.12"
langdetect==1.0.9 ; python_version == "3.12"
langgraph-checkpoint==2.0.25 ; python_version == "3.12"
langgraph-prebuilt==0.1.8 ; python_version == "3.12"
langgraph-sdk==0.1.64 ; python_version == "3.12"
langgraph==0.3.34 ; python_version == "3.12"
langsmith==0.3.38 ; python_version == "3.12"
lxml==5.4.0 ; python_version == "3.12"
markupsafe==3.0.2 ; python_version == "3.12"
marshmallow==3.26.1 ; python_version == "3.12"
matplotlib==3.10.1 ; python_version == "3.12"
multidict==6.4.3 ; python_version == "3.12"
mypy-extensions==1.1.0 ; python_version == "3.12"
narwhals==1.37.1 ; python_version == "3.12"
nest-asyncio==1.6.0 ; python_version == "3.12"
networkx==3.4.2 ; python_version == "3.12"
nltk==3.9.1 ; python_version == "3.12"
numpy==1.26.4 ; python_version == "3.12"
olefile==0.47 ; python_version == "3.12"
openai==1.76.2 ; python_version == "3.12"
openpyxl==3.1.5 ; python_version == "3.12"
opentelemetry-api==1.32.1 ; python_version == "3.12"
orjson==3.10.18 ; python_version == "3.12"
ormsgpack==1.9.1 ; python_version == "3.12"
packaging==24.2 ; python_version == "3.12"
pandas==2.2.3 ; python_version == "3.12"
pillow==11.2.1 ; python_version == "3.12"
propcache==0.3.1 ; python_version == "3.12"
proto-plus==1.26.1 ; python_version == "3.12"
protobuf==5.29.4 ; python_version == "3.12"
psutil==7.0.0 ; python_version == "3.12"
pyarrow==20.0.0 ; python_version == "3.12"
pyasn1-modules==0.4.2 ; python_version == "3.12"
pyasn1==0.6.1 ; python_version == "3.12"
pycparser==2.22 ; python_version == "3.12"
pydantic-core==2.33.2 ; python_version == "3.12"
pydantic-settings==2.9.1 ; python_version == "3.12"
pydantic==2.11.4 ; python_version == "3.12"
pydeck==0.9.1 ; python_version == "3.12"
pygraphviz==1.14 ; python_version == "3.12"
pyparsing==3.2.3 ; python_version == "3.12"
pypdf==5.4.0 ; python_version == "3.12"
python-dateutil==2.9.0.post0 ; python_version == "3.12"
python-docx==1.1.2 ; python_version == "3.12"
python-dotenv==1.1.0 ; python_version == "3.12"
python-iso639==2025.2.18 ; python_version == "3.12"
python-magic==0.4.27 ; python_version == "3.12"
python-oxmsg==0.0.2 ; python_version == "3.12"
python-pptx==1.0.2 ; python_version == "3.12"
pytz==2025.2 ; python_version == "3.12"
pyyaml==6.0.2 ; python_version == "3.12"
rapidfuzz==3.13.0 ; python_version == "3.12"
referencing==0.36.2 ; python_version == "3.12"
regex==2024.11.6 ; python_version == "3.12"
requests-toolbelt==1.0.0 ; python_version == "3.12"
requests==2.32.3 ; python_version == "3.12"
rpds-py==0.24.0 ; python_version == "3.12"
rsa==4.9.1 ; python_version == "3.12"
seaborn==0.13.2 ; python_version == "3.12"
six==1.17.0 ; python_version == "3.12"
smmap==5.0.2 ; python_version == "3.12"
sniffio==1.3.1 ; python_version == "3.12"
soupsieve==2.7 ; python_version == "3.12"
sqlalchemy==2.0.40 ; python_version == "3.12"
streamlit-code-editor==0.1.22 ; python_version == "3.12"
streamlit==1.45.0 ; python_version == "3.12"
tabulate==0.9.0 ; python_version == "3.12"
tenacity==9.1.2 ; python_version == "3.12"
tiktoken==0.9.0 ; python_version == "3.12"
toml==0.10.2 ; python_version == "3.12"
tornado==6.4.2 ; python_version == "3.12"
tqdm==4.67.1 ; python_version == "3.12"
typing-extensions==4.13.2 ; python_version == "3.12"
typing-inspect==0.9.0 ; python_version == "3.12"
typing-inspection==0.4.0 ; python_version == "3.12"
tzdata==2025.2 ; python_version == "3.12"
unstructured-client==0.34.0 ; python_version == "3.12"
unstructured==0.15.14 ; python_version == "3.12"
urllib3==2.4.0 ; python_version == "3.12"
watchdog==6.0.0 ; python_version == "3.12" and platform_system != "Darwin"
wrapt==1.17.2 ; python_version == "3.12"
xlrd==2.0.1 ; python_version == "3.12"
xlsxwriter==3.2.3 ; python_version == "3.12"
xxhash==3.5.0 ; python_version == "3.12"
yarl==1.20.0 ; python_version == "3.12"
zipp==3.21.0 ; python_version == "3.12"
zstandard==0.23.0 ; python_version == "3.12"
